added functionality for entity manager, some silly ball gameplay

+++ SFMLSimpleApp1/Engine/Math/Vector2.h
@@ -47,7 +47,7 @@ struct Vector2 {
 	}
 	float magnitude()
 	{
-		return sqrt(magnitudeSqr());
+		return sqrtf(magnitudeSqr());
 	}
 	void normalize()
 	{
